Rename REMOVE-PROPERTIES-xxx to DELETE-PROPERTIES-xxx
... after the realization, that this more consistent with the
way DELETE/REMOVE are used in the Lisp standard library (from
a user's point of view, DELETE-PROPERTY is actually destructive,
as it modifies an observable property (no pun intended) of the
object carrying the property list.)

The old names are still available, and probably will for some
time, since I will not get around to fixing all my own callers
of these functions. They should, however, not be used in new
code!