Introduce separate packages for topics
The major "topics" of this library now provide their APIs through
separate API packages whose only purpose is to expose the symbols
of the API in question. The definition of the symbols still lives
in the shared "tools" package.

This change preserves backward compatibility with code, which uses
the "big" package DARTS.LIB.TOOLS while allowing new code to access
only what it needs via topic package. In the long run, we should
get rid of the big package entirely, but that would require changing
too much code as of now.